TrainerImpl を tests/helpersへ移動
