+++ tests/helpers.py
@@ -9,6 +9,8 @@ from torch.utils.data import TensorDataset
 
 from ami.data.buffers.base_data_buffer import BaseDataBuffer
 from ami.data.step_data import DataKeys, StepData
+from ami.models.model_wrapper import ModelWrapper
+from ami.trainers.base_trainer import BaseTrainer
 
 
 def get_gpu_device() -> torch.device | None:
@@ -50,3 +52,22 @@ class DataBufferImpl(BaseDataBuffer):
 
 def skip_if_platform_is_not_linux():
     return pytest.mark.skipif(platform.system() != "Linux", reason="Platform is not linux.")
+
+
+class TrainerImpl(BaseTrainer):
+    def on_model_wrappers_dict_attached(self) -> None:
+        super().on_model_wrappers_dict_attached()
+
+        self.model1: ModelWrapper[ModelMultiplyP] = self.get_training_model("model1")
+        self.model2: ModelWrapper[ModelMultiplyP] = self.get_frozen_model("model2")
+
+    def on_data_users_dict_attached(self) -> None:
+        super().on_data_users_dict_attached()
+
+        self.data_user = self.get_data_user("buffer1")
+
+    def train(self) -> None:
+        dataset = self.data_user.get_new_dataset()
+        data = dataset[0][0]
+        self.model1(data)
+        self.model2(data)
