+++ tests/trainers/test_base_trainer.py
@@ -5,27 +5,7 @@ import torch.nn as nn
 from ami.data.utils import DataUsersDict
 from ami.models.model_wrapper import InferenceWrapper, ModelWrapper
 from ami.models.utils import ModelWrappersDict
-from ami.trainers.base_trainer import BaseTrainer
-from tests.helpers import ModelMultiplyP
-
-
-class TrainerImpl(BaseTrainer):
-    def on_model_wrappers_dict_attached(self) -> None:
-        super().on_model_wrappers_dict_attached()
-
-        self.model1: ModelWrapper[ModelMultiplyP] = self.get_training_model("model1")
-        self.model2: ModelWrapper[ModelMultiplyP] = self.get_frozen_model("model2")
-
-    def on_data_users_dict_attached(self) -> None:
-        super().on_data_users_dict_attached()
-
-        self.data_user = self.get_data_user("buffer1")
-
-    def train(self) -> None:
-        dataset = self.data_user.get_new_dataset()
-        data = dataset[0][0]
-        self.model1(data)
-        self.model2(data)
+from tests.helpers import ModelMultiplyP, TrainerImpl
 
 
 def assert_model_training(model: nn.Module) -> None:
